# docs/swagger.yaml
definitions:
  handler.AllOpeningsJSONReponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.OpeningResponse'
        type: array
    type: object
  handler.CreateOpening:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.OpeningJSONReponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpeningResponse'
    type: object
  handler.UpdateOpening:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  schemas.OpeningResponse:
    properties:
      company:
        type: string
      created_at:
        type: string
      id:
        type: integer
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /openings:
    get:
      consumes:
      - application/json
      description: get all job opening
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AllOpeningsJSONReponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: All Job Opening
      tags:
      - Openings
    post:
      consumes:
      - application/json
      description: Create a new job opening
      parameters:
      - description: Reques body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOpening'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.OpeningJSONReponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Opening
      tags:
      - Openings
  /openings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete job opening
      parameters:
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Opening
      tags:
      - Openings
    get:
      consumes:
      - application/json
      description: get job opening details
      parameters:
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OpeningJSONReponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Opening Details
      tags:
      - Openings
    put:
      consumes:
      - application/json
      description: Update a job opening
      parameters:
      - description: Reques body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOpening'
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Opening
      tags:
      - Openings
swagger: "2.0"